Implemented product Add and Edit page

+++ src/Shop/Shop.Application/Products/Edit/EditProductCommand.cs
@@ -6,46 +6,98 @@ using Common.Application.Utility.Validation.CustomFluentValidations;
 using FluentValidation;
 using Microsoft.AspNetCore.Http;
 using Shop.Application.Products._DTOs;
+using Shop.Domain.CategoryAggregate.Repository;
 using Shop.Domain.ProductAggregate;
 using Shop.Domain.ProductAggregate.Repository;
 using Shop.Domain.ProductAggregate.Services;
 
 namespace Shop.Application.Products.Edit;
 
-public record EditProductCommand(long ProductId, long CategoryId, string Name, string? EnglishName,
-    string Slug, string Description, IFormFile? MainImage, List<IFormFile>? GalleryImages,
-    List<ProductSpecificationDto>? CustomSpecifications,
-    List<ProductExtraDescriptionDto>? ExtraDescriptions) : IBaseCommand;
+public class EditProductCommand : IBaseCommand
+{
+    public long ProductId { get; set; }
+    public long CategoryId { get; set; }
+    public string Name { get; set; }
+    public string? EnglishName { get; set; }
+    public string Slug { get; set; }
+    public string? Introduction { get; set; }
+    public string? Review { get; set; }
+    public IFormFile? MainImage { get; set; }
+    public List<IFormFile>? GalleryImages { get; set; }
+    public List<ProductSpecificationDto>? Specifications { get; set; }
+    public List<ProductCategorySpecificationDto>? CategorySpecifications { get; set; }
+}
 
 public class EditProductCommandHandler : IBaseCommandHandler<EditProductCommand>
 {
     private readonly IProductRepository _productRepository;
-    private readonly IProductDomainService _productDomainService;
     private readonly IFileService _fileService;
+    private readonly IProductDomainService _productDomainService;
+    private readonly ICategoryRepository _categoryRepository;
 
-    public EditProductCommandHandler(IProductRepository productRepository,
-        IProductDomainService productDomainService, IFileService fileService)
+    public EditProductCommandHandler(IProductRepository productRepository, IFileService fileService,
+        IProductDomainService productDomainService, ICategoryRepository categoryRepository)
     {
         _productRepository = productRepository;
-        _productDomainService = productDomainService;
         _fileService = fileService;
+        _productDomainService = productDomainService;
+        _categoryRepository = categoryRepository;
     }
 
     public async Task<OperationResult> Handle(EditProductCommand request, CancellationToken cancellationToken)
     {
         var product = await _productRepository.GetAsTrackingAsync(request.ProductId);
-
         if (product == null)
-            return OperationResult.NotFound();
+            return OperationResult.NotFound(ValidationMessages.FieldNotFound("محصول"));
 
-        product.Edit(request.CategoryId, request.Name, request.EnglishName, request.Slug, request.Description,
-            _productDomainService);
+        var category = await _categoryRepository.GetAsTrackingAsync(request.CategoryId);
+        if (category == null)
+            return OperationResult.NotFound(ValidationMessages.FieldNotFound("دسته‌بندی"));
+
+        if (category.Specifications.Any()
+            && (request.CategorySpecifications == null || !request.CategorySpecifications.Any()))
+            return OperationResult.Error(ValidationMessages.CategorySpecificationRequired);
+        
+        if (request.CategorySpecifications != null && request.CategorySpecifications.Any())
+        {
+            var emptyOptionalSpecifications = new List<ProductCategorySpecificationDto>();
+
+            foreach (var spec in request.CategorySpecifications)
+            {
+                var categorySpec = category.Specifications.FirstOrDefault(s => s.Id == spec.CategorySpecificationId);
+
+                if (categorySpec == null)
+                    return OperationResult.Error("مشخصه دسته‌بندی یافت نشد");
+
+                if (categorySpec.IsOptional == false && string.IsNullOrWhiteSpace(spec.Description))
+                    return OperationResult.Error("لطفا مشخصات اجباری را وارد کنید");
+
+                if (categorySpec.IsOptional && string.IsNullOrWhiteSpace(spec.Description))
+                    emptyOptionalSpecifications.Add(spec);
+            }
+
+            if (emptyOptionalSpecifications.Count > 0)
+                emptyOptionalSpecifications.ForEach(spec =>
+                {
+                    request.CategorySpecifications?.Remove(spec);
+                });
+
+            var categorySpecifications = new List<ProductCategorySpecification>();
+
+            request.CategorySpecifications.ToList().ForEach(specification =>
+                categorySpecifications.Add(new ProductCategorySpecification(product.Id, specification.CategorySpecificationId, specification.Description)));
+
+            product.SetCategorySpecifications(categorySpecifications);
+        }
+
+        product.Edit(request.CategoryId, request.Name, request.EnglishName, request.Slug, request.Introduction,
+            request.Review, _productDomainService);
 
         var oldImage = product.MainImage;
 
         if (request.MainImage != null)
         {
-            _fileService.DeleteFile(Directories.ProductMainImages, oldImage.Name);
+            _fileService.DeleteFile(Directories.ProductMainImages, oldImage);
 
             var newMainImage =
                 await _fileService.SaveFileAndGenerateName(request.MainImage, Directories.ProductMainImages);
@@ -53,39 +105,28 @@ public class EditProductCommandHandler : IBaseCommandHandler<EditProductCommand>
             product.SetMainImage(newMainImage);
         }
 
-        var oldGalleryImages = product.GalleryImages.Select(img => img.Name).ToList();
-
-        if (request.GalleryImages != null)
+        if (request.GalleryImages != null && request.GalleryImages.Any())
         {
+            var oldGalleryImages = product.GalleryImages.Select(img => img.Name).ToList();
+
             _fileService.DeleteMultipleFiles(Directories.ProductGalleryImages, oldGalleryImages);
 
-            var newGalleryImages =
+            var galleryImagesNames =
                 await _fileService.SaveMultipleFilesAndGenerateNames(request.GalleryImages,
                     Directories.ProductGalleryImages);
 
-            product.SetGalleryImages(newGalleryImages);
+            product.SetGalleryImages(galleryImagesNames);
         }
 
-        if (request.CustomSpecifications != null)
+        if (request.Specifications != null && request.Specifications.Any())
         {
             var customSpecifications = new List<ProductSpecification>();
 
-            request.CustomSpecifications.ToList().ForEach(specification =>
+            request.Specifications.ToList().ForEach(specification =>
                 customSpecifications.Add(new ProductSpecification(product.Id, specification.Title,
                     specification.Description)));
 
-            product.SetCustomSpecifications(customSpecifications);
-        }
-
-        if (request.ExtraDescriptions != null)
-        {
-            var extraDescriptions = new List<ProductExtraDescription>();
-
-            request.ExtraDescriptions.ToList().ForEach(extraDescription =>
-                extraDescriptions.Add(new ProductExtraDescription(product.Id, extraDescription.Title,
-                    extraDescription.Description)));
-
-            product.SetExtraDescriptions(extraDescriptions);
+            product.SetSpecifications(customSpecifications);
         }
 
         await _productRepository.SaveAsync();
@@ -97,57 +138,53 @@ public class EditProductCommandValidator : AbstractValidator<EditProductCommand>
 {
     public EditProductCommandValidator()
     {
+        RuleFor(p => p.ProductId)
+            .NotEmpty().WithMessage(ValidationMessages.ProductIdRequired);
+
+        RuleFor(p => p.CategoryId)
+            .NotEmpty().WithMessage(ValidationMessages.CategoryIdRequired);
+
         RuleFor(p => p.Name)
-            .NotNull().WithMessage(ValidationMessages.ProductNameRequired)
             .NotEmpty().WithMessage(ValidationMessages.ProductNameRequired)
             .MaximumLength(2000).WithMessage(ValidationMessages.FieldCharactersMaxLength("نام محصول", 50));
 
-        RuleFor(p => p.Name)
+        RuleFor(p => p.EnglishName)
             .MaximumLength(2000).WithMessage(ValidationMessages.FieldCharactersMaxLength("نام انگلیسی محصول", 50));
 
         RuleFor(p => p.Slug)
-            .NotNull().WithMessage(ValidationMessages.SlugRequired)
             .NotEmpty().WithMessage(ValidationMessages.SlugRequired)
             .MaximumLength(100).WithMessage(ValidationMessages.FieldCharactersMaxLength("اسلاگ", 100));
 
-        RuleFor(p => p.MainImage)
-            .NotNull().WithMessage(ValidationMessages.FieldRequired("عکس اصلی محصول"))
-            .NotEmpty().WithMessage(ValidationMessages.FieldRequired("عکس اصلی محصول"))
-            .JustImageFile();
+        RuleFor(p => p.Introduction)
+            .MaximumLength(2000).WithMessage(ValidationMessages.FieldCharactersMaxLength("معرفی", 2000));
 
-        RuleFor(p => p.GalleryImages)
-            .NotNull().WithMessage(ValidationMessages.FieldRequired("عکس های گالری"))
-            .NotEmpty().WithMessage(ValidationMessages.FieldRequired("عکس های گالری"));
+        RuleFor(p => p.Review)
+            .MaximumLength(10000).WithMessage(ValidationMessages.FieldCharactersMaxLength("بررسی تخصصی", 10000));
+
+        RuleFor(p => p.MainImage)
+            .JustImageFile();
 
         RuleForEach(p => p.GalleryImages)
-            .NotNull().WithMessage(ValidationMessages.FieldRequired("عکس های گالری"))
-            .NotEmpty().WithMessage(ValidationMessages.FieldRequired("عکس های گالری"))
             .JustImageFile();
 
-        RuleForEach(p => p.CustomSpecifications).ChildRules(specification =>
+        RuleForEach(p => p.Specifications).ChildRules(specification =>
         {
             specification.RuleFor(spec => spec.Title)
-                .NotNull().WithMessage(ValidationMessages.TitleRequired)
                 .NotEmpty().WithMessage(ValidationMessages.TitleRequired)
                 .MaximumLength(100).WithMessage(ValidationMessages.FieldCharactersMaxLength("عنوان", 100));
 
             specification.RuleFor(spec => spec.Description)
-                .NotNull().WithMessage(ValidationMessages.TitleRequired)
-                .NotEmpty().WithMessage(ValidationMessages.TitleRequired)
+                .NotEmpty().WithMessage(ValidationMessages.DescriptionRequired)
                 .MaximumLength(300).WithMessage(ValidationMessages.FieldCharactersMaxLength("توضیحات", 300));
         });
 
-        RuleForEach(p => p.ExtraDescriptions).ChildRules(specification =>
+        RuleForEach(p => p.CategorySpecifications).ChildRules(categorySpecification =>
         {
-            specification.RuleFor(description => description.Title)
-                .NotNull().WithMessage(ValidationMessages.TitleRequired)
-                .NotEmpty().WithMessage(ValidationMessages.TitleRequired)
-                .MaximumLength(100).WithMessage(ValidationMessages.FieldCharactersMaxLength("عنوان", 100));
+            categorySpecification.RuleFor(p => p.CategorySpecificationId)
+                .NotEmpty().WithMessage(ValidationMessages.CategorySpecificationIdRequired);
 
-            specification.RuleFor(description => description.Description)
-                .NotNull().WithMessage(ValidationMessages.DescriptionRequired)
-                .NotEmpty().WithMessage(ValidationMessages.DescriptionRequired)
-                .MaximumLength(2000).WithMessage(ValidationMessages.FieldCharactersMaxLength("توضیحات", 2000));
+            categorySpecification.RuleFor(spec => spec.Description)
+                .MaximumLength(300).WithMessage(ValidationMessages.FieldCharactersMaxLength("توضیحات", 300));
         });
     }
 }
